fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,13 +6,13 @@
  */
 int main(void)
 {
-	int x = 1, y = 1, z = 1, a;
+	long int x = 1, y = 1, z = 1, a;
 
-	printf("%d", y);
+	printf("%ld", y);
 	while (x < 50)
 	{
 		a = y + z;
-		printf(", %d", a);
+		printf(", %ld", a);
 		y = z;
 		z = a;
 		x++;
